Extraigo la lista de series destacadas a otro componente
    refs:#168
